refactor(mm-next): move logAxiosError
