+++ packages/mirror-media-next/pages/story/[slug].js
@@ -22,9 +22,9 @@ import {
   getResizedUrl,
   getCategoryOfWineSlug,
   getLogTraceObject,
-  logAxiosError,
   logGqlError,
 } from '../../utils'
+import { logAxiosError } from '../../utils/log/shared'
 import { handleStoryPageRedirect } from '../../utils/story'
 import { MirrorMedia } from '@mirrormedia/lilith-draft-renderer'
 import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
